Revert "ref: Simplify exam status handling before polling"
This reverts commit b65ad5641e810f746602fb5951b68f5139c38e0b.

The reverted commit introduced a bug where the state 'EVALUATING' was not
handled correctly.
Thisone fixes it rolling back the logit as it was befor that change.

How can we call that? A reverse typo? Let's just say I brainfarted: I
thought that the second `else if` was against 'EVALUATING'.
Too many hours coding, I'd say.

Thanks @nishanb for the catch! Updated accordingly.